+++ src/main/java/co/com/park/gp/dto/LoginDTO.java
@@ -6,17 +6,17 @@ import co.com.park.gp.crosscutting.helpers.TextHelper;
 public final class LoginDTO {
 
     private TipoEmpleadoDTO tipoEmpleado;
-    private int numeroIdentificacion;
+    private Long numeroIdentificacion;
     private String password;
 
     public LoginDTO() {
         super();
         setTipoEmpleado(TipoEmpleadoDTO.build());
-        setNumeroIdentificacion(0);
+        setNumeroIdentificacion(0L);
         setPassword(TextHelper.EMPTY);
     }
 
-    public LoginDTO(TipoEmpleadoDTO tipoEmpleado, String password, int numeroIdentificacion) {
+    public LoginDTO(TipoEmpleadoDTO tipoEmpleado, String password, Long numeroIdentificacion) {
         setTipoEmpleado(tipoEmpleado);
         setNumeroIdentificacion(numeroIdentificacion);
         setPassword(password);
@@ -31,7 +31,7 @@ public final class LoginDTO {
         return this;
     }
 
-    public LoginDTO setNumeroIdentificacion(final int numeroIdentificacion) {
+    public LoginDTO setNumeroIdentificacion(final Long numeroIdentificacion) {
         this.numeroIdentificacion = numeroIdentificacion;
         return this;
     }
@@ -45,7 +45,7 @@ public final class LoginDTO {
         return tipoEmpleado;
     }
 
-    public int getNumeroIdentificacion() {
+    public Long getNumeroIdentificacion() {
         return numeroIdentificacion;
     }
